feat: add basic folders atomic designe
