+++ lambda-ui-graphql-relay-ts/.vscode/settings.json
@@ -1,0 +1,38 @@
+{
+  "flow.useNPMPackagedFlow": true,
+  "javascript.validate.enable": false,
+  "typescript.validate.enable": false,
+  "search.exclude": {
+    "**/node_modules": true,
+    "**/bower_components": true,
+    "**/build": true,
+    "**/flow-typed": true,
+    "**/scripts": true
+  },
+  "editor.formatOnSave": true,
+  "javascript.format.enable": false,
+  //"beautify.ignore": ["**/*.js", "**/*.jsx"],
+  "workbench.colorCustomizations": {
+    "activityBar.activeBackground": "#011f38",
+    "activityBar.activeBorder": "#bf001d",
+    "activityBar.background": "#011f38",
+    "activityBar.foreground": "#ebeff3",
+    "activityBar.inactiveForeground": "#b4b8bc99",
+    "activityBarBadge.background": "#bf001d",
+    "activityBarBadge.foreground": "#e7e7e7",
+    "statusBar.background": "#011526",
+    "statusBar.foreground": "#e7e7e7",
+    "statusBarItem.hoverBackground": "#011f38",
+    "titleBar.activeBackground": "#011526",
+    "titleBar.activeForeground": "#e7e7e7",
+    "titleBar.inactiveBackground": "#007fff99",
+    "titleBar.inactiveForeground": "#e7e7e799",
+    "sash.hoverBorder": "#011f38",
+    "statusBarItem.remoteBackground": "#011526",
+    "statusBarItem.remoteForeground": "#e7e7e7"
+  },
+  "peacock.color": "#007fff",
+  "workbench.editor.enablePreview": false,
+  "workbench.editor.enablePreviewFromQuickOpen": false,
+  "typescript.disableAutomaticTypeAcquisition": true
+}
