+++ lambda-ui-graphql-relay-ts/src/App.tsx
@@ -2,23 +2,14 @@ import "./App.css";
 
 import * as React from "react";
 
-import useConfig from "./components/useConfig";
-import logo from "./logo.svg";
+import Home from "./pages/Home.component";
 
 /**
  * Our Web Application
  */
+
 export default function App() {
-  const config = useConfig();
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <h1 className="App-title">Welcome to {config.app.TITLE}</h1>
-      </header>
-      <p className="App-intro">
-        To get started, edit <code>src/App.tsx</code> and save to reload.
-      </p>
-    </div>
+    <Home></Home>
   );
 }
