+++ lambda-ui-graphql-relay-ts/src/components/atoms/buttons/optional/Button2.component.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Button2() {
+    return (
+        <div>
+            <p>Boton opcional</p>
+        </div>
+    )
+}
+
+export default Button2
