+++ lambda-ui-graphql-relay-ts/src/components/atoms/buttons/standard/StandardButton.component.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function ButtonStandard() {
+    return (
+        <div>
+            <p>Hola</p>
+            <button>Hola</button>
+        </div>
+    )
+}
+
+export default ButtonStandard
