+++ lambda-ui-graphql-relay-ts/src/components/atoms/index.ts
@@ -1,0 +1,6 @@
+// Buttons
+import StandardButton from "./buttons/standard/StandardButton.component";
+import Button2 from "./buttons/optional/Button2.component";
+
+// Others
+export { StandardButton, Button2 };
